added some code fixes

+++ src/loader_lambda.py
@@ -31,10 +31,8 @@ def lambda_handler(event, context):
         3- registers the run number.
     """
     try:
-        if not os.path.isdir('./tmp'):
-            os.makedirs('./tmp')
-        else:
-            clean_tmp() 
+        create_dirs()
+
         # download .parquet files to pqt_tmp folder
         parquet_bucket = s3_list_prefix_parquet_buckets()
         dowload_parquet_files_to_process(parquet_bucket)
@@ -59,18 +57,36 @@ def lambda_handler(event, context):
 # S3 helper functions
 #####
 
+def create_dirs():
+    clean_tmp()
+
+    if not os.path.isdir('./tmp'):
+        os.makedirs('./tmp')
+    if not os.path.isdir('./tmp/csv_input'):
+        os.makedirs('./tmp/csv_input')
+    if not os.path.isdir('./tmp/csv_processed'):
+        os.makedirs('./tmp/csv_processed')
+    if not os.path.isdir('./tmp/pqt_input'):
+        os.makedirs('./tmp/pqt_input')
+    if not os.path.isdir('./tmp/pqt_processed'):
+        os.makedirs('./tmp/pqt_processed/')
+    
 def clean_tmp():
-    folders = subfolders = [ f.path for f in os.scandir('./tmp') if f.is_dir() ]
+    folders = [ f.path for f in os.scandir('./tmp') if f.is_dir() ]
+
     for folder in folders:
         files = os.listdir(folder)
         for file in files:
             os.remove(f'{folder}{file}')
-        os.removedirs(folder)
 
-    files = os.listdir('./tmp')
-    for file in files:
-        os.remove(f'./tmp/{file}')
-    # os.removedirs('./tmp')
+        os.rmdir(folder) # os.removedirs removes parent directory if it is empty in the last run!
+ 
+    if os.path.isdir('./tmp'):
+
+        files = os.listdir('./tmp')
+        for file in files:
+            os.remove(f'./tmp/{file}')
+
 
 def s3_list_buckets():
     """Using the s3 client, we will return a list containing the names of all the buckets.
@@ -153,7 +169,7 @@ def dowload_parquet_files_to_process(bucket):
         list=s3.list_objects(Bucket=bucket)['Contents']
         for key in list:
             if('input_parquet_key' in key['Key']):
-                s3.download_file(bucket, key['Key'], f'./pqt_tmp/{key["Key"].split("/")[1]}')
+                s3.download_file(bucket, key['Key'], f'./tmp/pqt_input/{key["Key"].split("/")[1]}')
     else:
         raise ValueError("ERROR: Terraform deployment unsuccessful")
     
@@ -194,17 +210,17 @@ def s3_create_parquet_processed_completed_txt_file(bucket):
     s3=boto3.client('s3')
 
     try:
-        s3.download_file(bucket, 'processed_parquet_key/parquet_processed.txt', './pqt_tmp/parquet_processed.txt')
+        s3.download_file(bucket, 'processed_parquet_key/parquet_processed.txt', './tmp/pqt_input/parquet_processed.txt')
     except Exception as error:
             raise ValueError(f'ERROR: {error}')
 
-    contents = open('./pqt_tmp/parquet_processed.txt', 'r').read()
+    contents = open('./tmp/pqt_input/parquet_processed.txt', 'r').read()
     num = int(contents.split(' ')[1])
-    with open('./pqt_tmp/parquet_processed.txt', 'w+') as file:
+    with open('./tmp/pqt_input/parquet_processed.txt', 'w+') as file:
         file.write(f'Run {num+1}')
     
     try:
-        s3.upload_file("./pqt_tmp/parquet_processed.txt", bucket, "processed_parquet_key/parquet_processed.txt")
+        s3.upload_file("./tmp/pqt_input/parquet_processed.txt", bucket, "processed_parquet_key/parquet_processed.txt")
     except Exception as error:
             raise ValueError(f'ERROR: {error}')
 
@@ -293,23 +309,23 @@ def fill_tables():
     engine = sa.create_engine(f'postgresql://{user}:{password}@{host}:5432/{database}')
 
 
-    dim_date = pd.read_parquet('./pqt_tmp/dim_date.parquet')
+    dim_date = pd.read_parquet('./tmp/pqt_input/dim_date.parquet')
     dim_date.to_sql('dim_date', engine, if_exists='append', index=False)
 
-    dim_design = pd.read_parquet('./pqt_tmp/dim_design.parquet')
+    dim_design = pd.read_parquet('./tmp/pqt_input/dim_design.parquet')
     dim_design.to_sql('dim_design', engine, if_exists='append', index=False)
     
-    dim_counterparty = pd.read_parquet('./pqt_tmp/dim_counterparty.parquet')
+    dim_counterparty = pd.read_parquet('./tmp/pqt_input/dim_counterparty.parquet')
     dim_counterparty.to_sql('dim_counterparty', engine, if_exists='append', index=False)
     
-    dim_currency = pd.read_parquet('./pqt_tmp/dim_currency.parquet')
+    dim_currency = pd.read_parquet('./tmp/pqt_input/dim_currency.parquet')
     dim_currency.to_sql('dim_currency', engine, if_exists='append', index=False)
 
-    dim_location = pd.read_parquet('./pqt_tmp/dim_location.parquet')
+    dim_location = pd.read_parquet('./tmp/pqt_input/dim_location.parquet')
     dim_location.to_sql('dim_location', engine, if_exists='append', index=False)
 
-    dim_staff = pd.read_parquet('./pqt_tmp/dim_staff.parquet')
+    dim_staff = pd.read_parquet('./tmp/pqt_input/dim_staff.parquet')
     dim_staff.to_sql('dim_staff', engine, if_exists='append', index=False)
 
-    fact_sales_order = pd.read_parquet('./pqt_tmp/fact_sales_order.parquet')
+    fact_sales_order = pd.read_parquet('./tmp/pqt_input/fact_sales_order.parquet')
     fact_sales_order.to_sql('fact_sales_order', engine, if_exists='append', index=False)
